requirements.txt
openai>=1.6.1,<2.0.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.1
python-dotenv==1.0.0
playwright==1.48.0
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
dnspython>=2.1.0
python-whois>=0.7.3